controller: make sh test grouping posix-compliant
-a and -o are not preferred in sh syntax:

http://wiki.bash-hackers.org/commands/classictest#and_and_or

-a and -o to mean AND and OR in a [ .. ] test expression is not well
defined. They are obsolescent extensions in POSIX and the lack of
additional quoting levels means that expressions involving these
operators are often ambiguous. Values that begin with hyphens and the !
string are the usual issues.

Using multiple [ .. ] expressions with shell AND/OR operators && and ||
is well defined and therefore preferred (but note that they have equal
precedence, while -a/-o is unspecified but usually implemented as -a
having higher precedence).

I have grouped together the AND statements for further clarification.